+++ src/app/(resources)/(feed)/atom.xml/route.ts
@@ -1,6 +1,6 @@
 'use server'
 
-import {getFeed} from '@/utils/feed'
+import {getFeed} from '@/lib/feed'
 
 export async function GET() {
   return new Response(getFeed().atom1(), {
